# Backend/Swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/email-template": {
      "get": {
        "operationId": "EmailTemplateController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EmailTemplate"
        ]
      },
      "post": {
        "operationId": "EmailTemplateController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmailDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "EmailTemplate"
        ]
      }
    },
    "/email-template/{id}": {
      "get": {
        "operationId": "EmailTemplateController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Email Template ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EmailTemplate"
        ]
      },
      "patch": {
        "operationId": "EmailTemplateController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Email Template ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EmailTemplate"
        ]
      },
      "delete": {
        "operationId": "EmailTemplateController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Email Template ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EmailTemplate"
        ]
      }
    },
...
    "/listing": {
      "get": {
        "operationId": "ListingController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Listing"
        ]
      },
      "post": {
        "operationId": "ListingController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListingSchemaDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Listing"
        ]
      }
    },
    "/listing/{id}": {
      "get": {
        "operationId": "ListingController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Listing Id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Listing"
        ]
      },
      "patch": {
        "operationId": "ListingController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Listing Id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateListingDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Listing"
        ]
      },
      "delete": {
        "operationId": "ListingController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Listing Id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Listing"
        ]
      }
    },
...
  },
  "info": {
    "title": "ExBuySell - Api Documentation",
    "description": "",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateEmailDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 4
          },
          "subject": {
            "type": "string",
            "minLength": 4
          },
          "cc": {
            "items": {
              "type": "array"
            },
            "nullable": true,
            "type": "array"
          },
          "body": {
            "type": "string",
            "minLength": 4
          }
        },
        "required": [
          "name",
          "subject",
          "body"
        ]
      },
      "UpdateEmailDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 4
          },
          "subject": {
            "type": "string",
            "minLength": 4
          },
          "cc": {
            "items": {
              "type": "array"
            },
            "nullable": true,
            "type": "array"
          },
          "body": {
            "type": "string",
            "minLength": 4
          }
        }
      },
...
      "ListingSchemaDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PUBLISH",
              "DRAFT"
            ]
          },
          "brand": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "minLength": 2,
                "required": false
              },
              "name": {
                "type": "string",
                "minLength": 2,
                "required": true
              },
              "domain": {
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 2
                },
                "minItems": 1,
                "required": true,
                "isArray": true
              },
              "business_location": {
                "type": "string",
                "minLength": 3,
                "required": true
              }
            }
          },
          "category": {
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "tools": {
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "financials": {
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "statistics": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "minLength": 2,
                "required": false
              },
              "conversion_rate": {
                "type": "string",
                "minLength": 2,
                "required": false
              },
              "customer_base": {
                "type": "string",
                "minLength": 2,
                "required": true
              },
              "average_order_value": {
                "type": "string",
                "minLength": 2,
                "required": false
              },
              "returning_customer": {
                "type": "string",
                "minLength": 2,
                "required": false
              },
              "email_subscribers": {
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 2
                },
                "required": false,
                "isArray": true
              },
              "refund_rate": {
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 2
                },
                "required": true,
                "isArray": true
              },
              "sales_channel": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "minLength": 2,
                      "required": false
                    },
                    "name": {
                      "type": "string",
                      "minLength": 2,
                      "required": true
                    },
                    "percentage": {
                      "type": "string",
                      "minLength": 2,
                      "required": true
                    }
                  },
                  "required": [
                    "name",
                    "percentage"
                  ]
                },
                "required": true,
                "isArray": true
              },
              "adverstising_channel": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "minLength": 2,
                      "required": false
                    },
                    "name": {
                      "type": "string",
                      "minLength": 2,
                      "required": true
                    },
                    "percentage": {
                      "type": "string",
                      "minLength": 2,
                      "required": true
                    }
                  },
                  "required": [
                    "name",
                    "percentage"
                  ]
                },
                "required": true,
                "isArray": true
              },
              "sales_countries": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "minLength": 2,
                      "required": false
                    },
                    "name": {
                      "type": "string",
                      "minLength": 2,
                      "required": true
                    },
                    "percentage": {
                      "type": "string",
                      "minLength": 2,
                      "required": true
                    }
                  },
                  "required": [
                    "name",
                    "percentage"
                  ]
                },
                "required": true,
                "isArray": true
              }
            }
          },
          "productQuestion": {
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "managementQuestion": {
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "social_account": {
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "advertisement": {
            "type": "object",
            "properties": {
              "price": {
                "type": "string",
                "minLength": 2,
                "required": true
              },
              "title": {
                "type": "string",
                "minLength": 2,
                "required": true
              },
              "intro": {
                "type": "string",
                "minLength": 2,
                "required": true
              },
              "usp": {
                "type": "string",
                "minLength": 2,
                "required": true
              },
              "description": {
                "type": "string",
                "minLength": 2,
                "required": true
              }
            }
          },
          "handover": {
            "type": "object",
            "properties": {
              "handover_option": {
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 2
                },
                "minItems": 1,
                "required": true,
                "isArray": true
              },
              "is_sale_support": {
                "type": "string",
                "minLength": 4,
                "maxLength": 4,
                "required": true
              },
              "total_month": {
                "type": "string",
                "minLength": 1,
                "required": true
              }
            }
          }
        },
        "required": [
          "status",
          "brand",
          "category",
          "tools",
          "financials",
          "statistics",
          "productQuestion",
          "managementQuestion",
          "social_account",
          "advertisement",
          "handover"
        ]
      },
      "UpdateListingDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PUBLISH",
              "DRAFT"
            ]
          },
          "brand": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "minLength": 2,
                "required": false
              },
              "name": {
                "type": "string",
                "minLength": 2,
                "required": true
              },
              "domain": {
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 2
                },
                "minItems": 1,
                "required": true,
                "isArray": true
              },
              "business_location": {
                "type": "string",
                "minLength": 3,
                "required": true
              }
            }
          },
          "category": {
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "tools": {
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "financials": {
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "statistics": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "minLength": 2,
                "required": false
              },
              "conversion_rate": {
                "type": "string",
                "minLength": 2,
                "required": false
              },
              "customer_base": {
                "type": "string",
                "minLength": 2,
                "required": true
              },
              "average_order_value": {
                "type": "string",
                "minLength": 2,
                "required": false
              },
              "returning_customer": {
                "type": "string",
                "minLength": 2,
                "required": false
              },
              "email_subscribers": {
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 2
                },
                "required": false,
                "isArray": true
              },
              "refund_rate": {
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 2
                },
                "required": true,
                "isArray": true
              },
              "sales_channel": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "minLength": 2,
                      "required": false
                    },
                    "name": {
                      "type": "string",
                      "minLength": 2,
                      "required": true
                    },
                    "percentage": {
                      "type": "string",
                      "minLength": 2,
                      "required": true
                    }
                  },
                  "required": [
                    "name",
                    "percentage"
                  ]
                },
                "required": true,
                "isArray": true
              },
              "adverstising_channel": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "minLength": 2,
                      "required": false
                    },
                    "name": {
                      "type": "string",
                      "minLength": 2,
                      "required": true
                    },
                    "percentage": {
                      "type": "string",
                      "minLength": 2,
                      "required": true
                    }
                  },
                  "required": [
                    "name",
                    "percentage"
                  ]
                },
                "required": true,
                "isArray": true
              },
              "sales_countries": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "minLength": 2,
                      "required": false
                    },
                    "name": {
                      "type": "string",
                      "minLength": 2,
                      "required": true
                    },
                    "percentage": {
                      "type": "string",
                      "minLength": 2,
                      "required": true
                    }
                  },
                  "required": [
                    "name",
                    "percentage"
                  ]
                },
                "required": true,
                "isArray": true
              }
            }
          },
          "productQuestion": {
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "managementQuestion": {
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "social_account": {
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "advertisement": {
            "type": "object",
            "properties": {
              "price": {
                "type": "string",
                "minLength": 2,
                "required": true
              },
              "title": {
                "type": "string",
                "minLength": 2,
                "required": true
              },
              "intro": {
                "type": "string",
                "minLength": 2,
                "required": true
              },
              "usp": {
                "type": "string",
                "minLength": 2,
                "required": true
              },
              "description": {
                "type": "string",
                "minLength": 2,
                "required": true
              }
            }
          },
          "handover": {
            "type": "object",
            "properties": {
              "handover_option": {
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 2
                },
                "minItems": 1,
                "required": true,
                "isArray": true
              },
              "is_sale_support": {
                "type": "string",
                "minLength": 4,
                "maxLength": 4,
                "required": true
              },
              "total_month": {
                "type": "string",
                "minLength": 1,
                "required": true
              }
            }
          }
        }
      },
...
    }
  }
}